Use NIOAsyncTestingChannel in tests (#197)
* Use NIOAsyncTestingChannel for testing

* Use sendConnect() in MQTTConnection.connect

* Update after merge

* Close is nonisolated but still sends a disconnect

* Add some helper functions

* Update integration tests

* Don't send two DISCONNECTs

* Fix up after rebase

* Fix mutable data warning

* Update Sources/MQTTNIO/Connection/MQTTConnection.swift


* Update Sources/MQTTNIO/Connection/MQTTConnection.swift


* Update Tests/IntegrationTests/IntegrationTests.swift


* Update Tests/IntegrationTests/IntegrationTests.swift


---------
